tests/wpt/mozilla/tests/mozilla/bluetooth/requestDevice/acceptAllDevices/accept-all-devices.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.glucose_heart_rate);
    return window.navigator.bluetooth.requestDevice({acceptAllDevices: true})
    .then(device => assert_true([mock_device_name.heart_rate, mock_device_name.glucose].includes(device.name)));
}, 'Requesting all devices without filters must return all the devices.');
</script>
